% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{col.headings}
\alias{col.headings}
\title{Specify the columns of the data frame required by score imputation method}
\usage{
col.headings(arm, has.event, time, Id, DCO.time, to.impute,
  censor.type = NULL)
}
\arguments{
\item{arm}{column name which will contain the subject's treatment group}

\item{has.event}{column name which will contain whether 
the subject has an event (1) or not(0)}

\item{time}{column name of censoring/event time}

\item{Id}{column name of subject Id}

\item{DCO.time}{column name of the time at which the subject would have been
censored had they not had an event before data cut off}

\item{to.impute}{column name of the logical column as to whether events should
be imputed}

\item{censor.type}{column name of the column containing the reason for censoring,
0=had event, 1=non-administrative censoring 2=administrative censoring -- only subjects
with 1 in this column count as having an `event' in the Cox model for censoring 
(optionally used -- if not used then all subjects who are censored are used)}
}
\value{
A list contain the given arguments
}
\description{
Specify the columns of the data frame required by score imputation method
}

