% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaImputeData.R, R/gammaStat.R, R/generics.R, R/scoreImputedSet.R, R/scoreStatSet.R
\name{ExtractSingle}
\alias{ExtractSingle}
\alias{ExtractSingle.GammaImputedSet}
\alias{ExtractSingle.GammaStatList}
\alias{ExtractSingle.ScoreImputedSet}
\alias{ExtractSingle.ScoreStatList}
\title{Extract a single risk score/gamma imputed data set/model fit}
\usage{
\method{ExtractSingle}{GammaImputedSet}(x, index)

\method{ExtractSingle}{GammaStatList}(x, index)

ExtractSingle(x, index)

\method{ExtractSingle}{ScoreImputedSet}(x, index)

\method{ExtractSingle}{ScoreStatList}(x, index)
}
\arguments{
\item{x}{The multiple imputed object}

\item{index}{Integer, which imputed data set/model fit should be returned}
}
\value{
The individual data set/model fit
}
\description{
Extract a single risk score/gamma imputed data set/model fit
}

