% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gammaImputeData.R
\name{GammaImputedData.object}
\alias{GammaImputedData.object}
\title{\code{GammaImputedData} object}
\description{
An object which contains
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing the time to event data
with 3 new columns impute.time and impute.event, the imputed event/censoring times and event indicators
(for subjects whose data is not imputed these columns contain the unchanged event/censoring time and
event indicator) and internal_gamma_val which is the value of gamma used for each subject in this data set}

\item{\code{default.formula}}{The default model formula which will be used when fitting the imputed data}
}}

