% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaImputeData.R
\name{GammaImputedSet.object}
\alias{GammaImputedSet.object}
\title{\code{GammaImputedSet} object}
\description{
An object which contains the set of gamma imputed data frames.
Use the \code{ExtractSingle} function to extract a single
\code{GammaImputedData} objects. Use the ImputeStat function to fit models
to the entire set of imputed data frames
}
\details{
It contains the following:
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing the unimputed time to event data (along with a column internal_gamma_val
which is the value of gamma used for each subject in this data set)}

\item{\code{m}}{The number of imputed data sets}

\item{\code{gamma.factor}}{The value of gamma.factor used with the imputation}

\item{\code{impute.time}}{A matrix (1 column per imputed data set) containing the imputed times}

\item{\code{impute.event}}{A matrix (1 column per imputed data set) containing the imputed event indicators}

\item{\code{default.formula}}{The default model formula which will be used when fitting the imputed data}
}}
\seealso{
\code{\link{GammaImputedData.object}}
}

