% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gammaStat.R
\name{GammaStat.object}
\alias{GammaStat.object}
\title{\code{GammaStat} object}
\description{
An S3 object which contains the point estimate
and test statistic after fitting a model to
a \code{GammaImputedData} object.
}
\details{
The function \code{print.GammaStat} has been implemented

The object contains the following:
}
\section{Slots}{

\describe{
\item{\code{model}}{The model used to create the fit}

\item{\code{method}}{The model used for the fit}

\item{\code{estimate}}{A point estimate of the test statistic}

\item{\code{var}}{The estimate for the variance of the test statistic}
}}

