% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gammaStat.R
\name{GammaStatList.object}
\alias{GammaStatList.object}
\title{\code{GammaStatList} object}
\description{
The object containing the results of fitting models to
a \code{GammaImputedSet} object.
}
\details{
A \code{summary.GammaStatList} has been implemented which performs
Rubin's multiple imputation rules.

The object contains the following
}
\section{Slots}{

\describe{
\item{\code{fits}}{A list of \code{GammaStat} objects containing the model fits for
the imputed data sets}

\item{\code{statistics}}{A list with two elements: estimates and vars which contain the coefficient
estimates and their variances one column per covariate one row per imputed data set}

\item{\code{m}}{The number of model fits}
}}

