% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\docType{package}
\name{InformativeCensoring-package}
\alias{InformativeCensoring}
\alias{InformativeCensoring-package}
\title{Perform methods of multiple imputation for
time to event data}
\description{
Perform methods of multiple imputation for
time to event data
}
\details{
See Nonparametric comparison of two survival functions with
dependent censoring via nonparametric multiple imputation. Hsu and Taylor
Statistics in Medicine (2009) 28:462-475 for Hsu's method

See Relaxing the independent censoring assumption in the Cox proportional
hazards model using multiple imputation. Jackson et al., Statistics in Medicine
(2014) 33:4681-4694 for Jackson's method
}
\author{
\email{David.Ruaa@astrazeneca.com}
}

