% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedependent.R
\name{MakeTimeDepScore}
\alias{MakeTimeDepScore}
\title{Create a valid \code{ScoreTD} object}
\usage{
MakeTimeDepScore(data, Id, time.start, time.end)
}
\arguments{
\item{data}{A data frame of time dependent covariates}

\item{Id}{The column name of the subject Id}

\item{time.start}{The covariates are valid for the time [time.start,time.end] where
time.start is the column name of time.start}

\item{time.end}{The covariates are valid for the time [time.start,time.end] where
time.end is the column name of time.end}
}
\value{
A \code{ScoreTD} object
}
\description{
Create a valid \code{ScoreTD} object
}

