% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{NN.options}
\alias{NN.options}
\title{Create a list of options which control the nearest neighbour algorithm
for risk score imputation}
\usage{
NN.options(NN = 5, w.censoring = 0.2, min.subjects = 20)
}
\arguments{
\item{NN}{The (maximum) number of subjects to be included in the
risk set}

\item{w.censoring}{The weighting on the censoring risk score when
calculating distances for the nearest neighbour calculation
A weighting of \code{(1-w.censoring)} is used for the event risk score}

\item{min.subjects}{If using time dependent score imputation include at least
this number of subjects when fitting the Cox model (i.e. include some subjects who were censored/had event
earlier than the cenosred observation if neccessary)}
}
\value{
A list of options used within the ScoreImputedData
function
}
\description{
Create a list of options which control the nearest neighbour algorithm
for risk score imputation
}

