% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreImputedData.R
\name{ScoreImputedData.object}
\alias{ScoreImputedData.object}
\title{\code{ScoreImputedData} object}
\description{
An object which contains
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing the time to event data
with 2 new columns impute.time and impute.event, the imputed event/censoring times and event indicators
(for subjects whose data is not imputed these columns contain the unchanged event/censoring time and
event indicator )}

\item{\code{col.control}}{The list of column names the risk score imputation method requires see \code{\link{col.headings}}
for further details. If censor.type was not used then \code{col.control$censor.type="using_has.event_col"}}

\item{\code{default.formula}}{The default model formula which will be used when fitting the imputed data using a Cox model}
}}

