% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreImputedSet.R
\name{ScoreImputedSet.object}
\alias{ScoreImputedSet.object}
\title{\code{ScoreImputedSet} object}
\description{
An object which contains the set of score imputed data frames.
Use the \code{ExtractSingle} function to extract a single 
\code{ScoreImputedData} object. Use the \code{ScoreStat} function to fit models
to the entire set of imputed data frames
}
\details{
It contains the following:
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing the unimputed time to event data}

\item{\code{col.control}}{The list of column names the score imputation method requires see \code{\link{col.headings}}
for further details}

\item{\code{m}}{The number of imputed data sets}

\item{\code{impute.time}}{A matrix (1 column per imputed data set) containing the imputed times}

\item{\code{impute.event}}{A matrix (1 column per imputed data set) containing the imputed event indicators}

\item{\code{default.formula}}{The default model formula which will be used when fitting the imputed data using a Cox model}
}}
\seealso{
\code{\link{ScoreImputedData.object}}
}

