% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreStat.R
\name{ScoreStat.object}
\alias{ScoreStat.object}
\title{ScoreStat object}
\description{
An S3 object which contains the point estimate 
and test statistic after fitting a model to 
a \code{ScoreImputedData} object.
}
\details{
The functions \code{print.ScoreStat} and \code{as.vector.ScoreStat}
have been included

The object contains the following:

The test statistic should be normally distributed and hence for
the logrank test Z = (O_2 - E_2)/sqrt(V_2), i.e. the square root of the standard
Chi squared statistic (with the appropriate sign)
}
\section{Slots}{

\describe{
\item{\code{model}}{The model used to create the fit}

\item{\code{method}}{The method used for the fit}

\item{\code{estimate}}{A point estimate of the test statistic}

\item{\code{var}}{The estimate for the variance of the test statistic}

\item{\code{statistic}}{The test statistic given by \code{estimate/sqrt(var)}}
}}

