% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreStatSet.R
\name{ScoreStatList.object}
\alias{ScoreStatList.object}
\title{ScoreStatList}
\description{
The object containing the results of fitting models to
a \code{ScoreImputedSet} object.
}
\details{
A \code{summary.ScoreStatList} has been implemented.

The object contains the following
}
\section{Slots}{

\describe{
\item{\code{fits}}{A list of \code{ScoreStat} objects containing the model fits for
the imputed data sets}

\item{\code{statistics}}{A \code{ScoreStatSet} object containing the statistics}

\item{\code{m}}{The number of model fits}
}}
\seealso{
\code{\link{ScoreStatSet.object}} \code{\link{ScoreStat.object}}
}

