% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreStatSet.R
\name{ScoreStatSet.object}
\alias{ScoreStatSet.object}
\title{An object which contains the test statistic and estimators for
a set of model fits to imputed data using risk score imputation}
\description{
The object is a Mx3 matrix, one row per imputed data set
and columns: estimate (the point estimates), var (their variances)
and Z (the test statistic). M must be > 4
}
\details{
Note the Z should be ~ standard normal (so we do not use the chi_squared
test statistic see \code{\link{ScoreStat.object}} for further details)

The summary.ScoreStatSet function will apply the MI averaging procedures
and estimates of the test statistic and p-value
}
\seealso{
\code{\link{summary.ScoreStatSet}}
}

