% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreStatSet.R
\name{ScoreStatSet}
\alias{ScoreStatSet}
\title{S3 generic to create a \code{ScoreStatSet} object}
\usage{
ScoreStatSet(x)
}
\arguments{
\item{x}{The object to convert into a \code{ScoreStatSet} object}
}
\value{
A ScoreStatSet object
}
\description{
S3 generic to create a \code{ScoreStatSet} object
}
\seealso{
\code{\link{ScoreStatSet.object}}
}

