% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timedependent.R
\name{ScoreTD.object}
\alias{ScoreTD.object}
\title{A \code{ScoreTD} object}
\description{
This data frame holds time dependent covariates for
use with risk score imputation
}
\details{
The data frame contains the following columns:
'Id' for subject ID \cr
'time.start' and 'time.end' the range of time for which
the covariate values are valid - i.e. [time.start,time.end] \cr
Additional columns are the time dependent covariates

All data for a single subject should be stored in consecutive rows, sorted
by time and the starting time of a row should match the ending time of the previous row
}
\seealso{
\code{\link{MakeTimeDepScore}}
}

