% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDescription.R
\docType{data}
\name{ScoreTimeDep}
\alias{ScoreTimeDep}
\title{Simulated time dependent variables for time to event data}
\format{A data.frame containing 1 row per subject-visit}
\description{
This data set contains time dependent covariates for the
\code{\link{ScoreInd}} time to event data.
}
\section{Fields}{

\describe{
\item{\code{Id}}{The Subject Id}

\item{\code{start}}{The covariate given in each row are for a given subject from time 'start'...}

\item{\code{end}}{... until time end}

\item{\code{W1}}{The value of a (binary) time dependent variable
for the subject with the given 'Id' between times 'start' and 'end'}

\item{\code{W2}}{The value of a (continuous) time dependent variable
for the subject with the given 'Id' between times 'start' and 'end'}
}}

