% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreStatSet.R
\name{summary.ScoreStatSet}
\alias{summary.ScoreStatSet}
\alias{summary.ScoreStatSet.object}
\title{Summary object of \code{ScoreStatSet} object}
\description{
This object contains the multiple imputed 
averages/p-values of a set of estimates from 
risk score imputed data sets.
}
\details{
A \code{print.summary.ScoreStatSet} function has been implemented

This object contains three lists meth1 and meth2 and methRubin
meth1 averages the point estimates to produce an F test statistic,
meth2 averages the test statistics and prodcues a t test statistic
and methRubin follows Rubin's standard rules and is used for calculating
confidence intervals

See the vignette for further details.

meth1, meth2 and methRubin are lists with the following elements:
estimate: average estimator for meth1, NOTE: for meth2 this is the average test statistic, \cr
var: estimate of variance of "estimate" field \cr
test.stat: test statistic \cr
distribution: distribution of statistical test (i.e. F or t) \cr
p.value: p-value of test \cr
}

